+++ docs/code_ref/instr.rst
@@ -4,6 +4,10 @@ SunPy instr
 The ``instr`` submodule contains instrument specific routines, generally this
 submodule is focused on routines that process calibrated data.
 
+.. warning::
+    This module will be removed in sunpy 3.1.
+    This code has been moved out to `sunkit-instruments <https://pypi.org/project/sunkit-instruments/>`__.
+
 
 .. automodapi:: sunpy.instr
