--- sunpy/data/test/hsi_obssumm_filedb_201104.txt
@@ -1,34 +1,0 @@
-HESSI Filedb File:
-Created: 2014-04-19T08:15:48.000
-Number of Files:           30
-                    Filename  Orb_st Orb_end         Start_time           End_time Status_flag    Npackets Drift_start   Drift_end Data source
-hsi_obssumm_20110401_043.fit       0       0 01-Apr-11 00:00:00 02-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110402_048.fit       0       0 02-Apr-11 00:00:00 03-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110403_048.fit       0       0 03-Apr-11 00:00:00 04-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110404_042.fit       0       0 04-Apr-11 00:00:00 05-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110405_031.fit       0       0 05-Apr-11 00:00:00 06-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110406_041.fit       0       0 06-Apr-11 00:00:00 07-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110407_031.fit       0       0 07-Apr-11 00:00:00 08-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110408_038.fit       0       0 08-Apr-11 00:00:00 09-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110409_030.fit       0       0 09-Apr-11 00:00:00 10-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110410_028.fit       0       0 10-Apr-11 00:00:00 11-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110411_024.fit       0       0 11-Apr-11 00:00:00 12-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110412_024.fit       0       0 12-Apr-11 00:00:00 13-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110413_022.fit       0       0 13-Apr-11 00:00:00 14-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110414_017.fit       0       0 14-Apr-11 00:00:00 15-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110415_020.fit       0       0 15-Apr-11 00:00:00 16-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110416_016.fit       0       0 16-Apr-11 00:00:00 17-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110417_017.fit       0       0 17-Apr-11 00:00:00 18-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110418_017.fit       0       0 18-Apr-11 00:00:00 19-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110419_024.fit       0       0 19-Apr-11 00:00:00 20-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110420_029.fit       0       0 20-Apr-11 00:00:00 21-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110421_025.fit       0       0 21-Apr-11 00:00:00 22-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110422_029.fit       0       0 22-Apr-11 00:00:00 23-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110423_030.fit       0       0 23-Apr-11 00:00:00 24-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110424_025.fit       0       0 24-Apr-11 00:00:00 25-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110425_026.fit       0       0 25-Apr-11 00:00:00 26-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110426_022.fit       0       0 26-Apr-11 00:00:00 27-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110427_029.fit       0       0 27-Apr-11 00:00:00 28-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110428_021.fit       0       0 28-Apr-11 00:00:00 29-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110429_028.fit       0       0 29-Apr-11 00:00:00 30-Apr-11 00:00:00           0           0       0.000       0.000                                                                                
-hsi_obssumm_20110430_029.fit       0       0 30-Apr-11 00:00:00 01-May-11 00:00:00           0           0       0.000       0.000
