+++ sunpy/timeseries/sources/rhessi.py
@@ -7,17 +7,92 @@ from collections import OrderedDict
 
 import matplotlib.dates
 import matplotlib.pyplot as plt
+import numpy as np
 from pandas import DataFrame
 
 import astropy.units as u
+from astropy.time import TimeDelta
 
 import sunpy.io
-from sunpy.instr import rhessi
+from sunpy.time import parse_time
 from sunpy.timeseries.timeseriesbase import GenericTimeSeries
 from sunpy.util.metadata import MetaDict
 from sunpy.visualization import peek_show
 
 __all__ = ['RHESSISummaryTimeSeries']
+
+
+def uncompress_countrate(compressed_countrate):
+    """
+    Convert the compressed count rate inside of observing summary file from a
+    compressed byte to a true count rate.
+
+    Parameters
+    ----------
+    compressed_countrate : `byte` array
+        A compressed count rate returned from an observing summary file.
+
+    References
+    ----------
+    `Hsi_obs_summ_decompress.pro <https://hesperia.gsfc.nasa.gov/ssw/hessi/idl/qlook_archive/hsi_obs_summ_decompress.pro>`_
+    """
+
+    # Ensure uncompressed counts are between 0 and 255
+    if (compressed_countrate.min() < 0) or (compressed_countrate.max() > 255):
+        raise ValueError(
+            f'Exepected uncompressed counts {compressed_countrate} to in range 0-255')
+
+    # TODO Must be a better way than creating entire lookup table on each call
+    ll = np.arange(0, 16, 1)
+    lkup = np.zeros(256, dtype='int')
+    _sum = 0
+    for i in range(0, 16):
+        lkup[16 * i:16 * (i + 1)] = ll * 2 ** i + _sum
+        if i < 15:
+            _sum = lkup[16 * (i + 1) - 1] + 2 ** i
+
+    return lkup[compressed_countrate]
+
+
+def parse_observing_summary_hdulist(hdulist):
+    """
+    Parse a RHESSI observation summary file.
+
+    Parameters
+    ----------
+    hdulist : `list`
+        The HDU list from the fits file.
+
+    Returns
+    -------
+    out : `dict`
+        Returns a dictionary.
+    """
+    header = hdulist[0].header
+
+    reference_time_ut = parse_time(hdulist[5].data.field('UT_REF')[0],
+                                   format='utime')
+    time_interval_sec = hdulist[5].data.field('TIME_INTV')[0]
+    # label_unit = fits[5].data.field('DIM1_UNIT')[0]
+    # labels = fits[5].data.field('DIM1_IDS')
+    labels = ['3 - 6 keV', '6 - 12 keV', '12 - 25 keV', '25 - 50 keV',
+              '50 - 100 keV', '100 - 300 keV', '300 - 800 keV',
+              '800 - 7000 keV', '7000 - 20000 keV']
+
+    # The data stored in the fits file are "compressed" countrates stored as
+    # one byte
+    compressed_countrate = np.array(hdulist[6].data.field('countrate'))
+
+    countrate = uncompress_countrate(compressed_countrate)
+    dim = np.array(countrate[:, 0]).size
+
+    time_array = parse_time(reference_time_ut) + \
+        TimeDelta(time_interval_sec * np.arange(dim) * u.second)
+
+    #  TODO generate the labels for the dict automatically from labels
+    data = {'time': time_array, 'data': countrate, 'labels': labels}
+
+    return header, data
 
 
 class RHESSISummaryTimeSeries(GenericTimeSeries):
@@ -139,7 +214,7 @@ class RHESSISummaryTimeSeries(GenericTimeSeries):
         hdulist : `astropy.io.fits.HDUList`
             A HDU list.
         """
-        header, d = rhessi.parse_observing_summary_hdulist(hdulist)
+        header, d = parse_observing_summary_hdulist(hdulist)
         # The time of dict `d` is astropy.time, but dataframe can only take datetime
         d['time'] = d['time'].datetime
         header = MetaDict(OrderedDict(header))
